move and rename useTimeZone because it's not a hook
